remove references to gizmo and asm consolidate packages
